refactor) PaymentDto 클래스 오류 수정
