+++ pay/src/main/java/com/zerobase/service/PayManagementService.java
@@ -8,10 +8,10 @@ import com.zerobase.api.TravelApi;
 import com.zerobase.entity.DepositEntity;
 import com.zerobase.entity.PaymentEntity;
 import com.zerobase.model.DepositDto;
+import com.zerobase.model.PaymentDto;
 import com.zerobase.model.ResponseApi;
 import com.zerobase.model.ResponseDepositPayDto;
 import com.zerobase.model.type.PGMethod;
-import com.zerobase.model.PaymentDto.PaymentDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
