+++ pay/src/test/java/com/zerobase/service/PaymentServiceTest.java
@@ -10,7 +10,7 @@ import static org.mockito.Mockito.when;
 import com.zerobase.entity.PaymentEntity;
 import com.zerobase.model.type.OrderType;
 import com.zerobase.model.type.PGMethod;
-import com.zerobase.model.PaymentDto.PaymentDto;
+import com.zerobase.model.PaymentDto;
 import com.zerobase.model.type.PaymentStatus;
 import com.zerobase.repository.PaymentRepository;
 import org.junit.jupiter.api.Test;
